Data.Profunctor.Indexed

Description

Definitions of concrete profunctors and profunctor classes.

Profunctor classes

class Profunctor p where

Minimal complete definition

dimap, lmap, rmap

Methods

dimap :: (a -> b) -> (c -> d) -> p i b c -> p i a d

lmap :: (a -> b) -> p i b c -> p i a c

rmap :: (c -> d) -> p i b c -> p i b d

lcoerce' :: Coercible a b => p i a c -> p i b c

default lcoerce' :: Coercible (p i a c) (p i b c) => p i a c -> p i b c

rcoerce' :: Coercible a b => p i c a -> p i c b

default rcoerce' :: Coercible (p i c a) (p i c b) => p i c a -> p i c b

conjoined__ :: (p i a b -> p i s t) -> (p i a b -> p j s t) -> p i a b -> p j s t

default conjoined__ :: Coercible (p i s t) (p j s t) => (p i a b -> p i s t) -> (p i a b -> p j s t) -> p i a b -> p j s t

ixcontramap :: (j -> i) -> p i a b -> p j a b

default ixcontramap :: Coercible (p i a b) (p j a b) => (j -> i) -> p i a b -> p j a b

lcoerce :: (Coercible a b, Profunctor p) => p i a c -> p i b c

lcoerce' with type arguments rearranged for TypeApplications.

rcoerce :: (Coercible a b, Profunctor p) => p i c a -> p i c b

rcoerce' with type arguments rearranged for TypeApplications.

class Profunctor p => Strong p where

Minimal complete definition

first', second'

Methods

first' :: p i a b -> p i (a, c) (b, c)

second' :: p i a b -> p i (c, a) (c, b)

linear :: (forall f. Functor f => (a -> f b) -> s -> f t) -> p i a b -> p i s t

ilinear :: (forall f. Functor f => (i -> a -> f b) -> s -> f t) -> p j a b -> p (i -> j) s t

default ilinear :: Coercible (p j s t) (p (i -> j) s t) => (forall f. Functor f => (i -> a -> f b) -> s -> f t) -> p j a b -> p (i -> j) s t

class Profunctor p => Costrong p where

Methods

unfirst :: p i (a, d) (b, d) -> p i a b

unsecond :: p i (d, a) (d, b) -> p i a b

class Profunctor p => Choice p where

Methods

left' :: p i a b -> p i (Either a c) (Either b c)

right' :: p i a b -> p i (Either c a) (Either c b)

class Profunctor p => Cochoice p where

Methods

unleft :: p i (Either a d) (Either b d) -> p i a b

unright :: p i (Either d a) (Either d b) -> p i a b

class (Choice p, Strong p) => Visiting p where

Minimal complete definition

Nothing

Methods

visit :: forall i s t a b. (forall f. Functor f => (forall r. r -> f r) -> (a -> f b) -> s -> f t) -> p i a b -> p i s t

ivisit :: (forall f. Functor f => (forall r. r -> f r) -> (i -> a -> f b) -> s -> f t) -> p j a b -> p (i -> j) s t

default ivisit :: Coercible (p j s t) (p (i -> j) s t) => (forall f. Functor f => (forall r. r -> f r) -> (i -> a -> f b) -> s -> f t) -> p j a b -> p (i -> j) s t

class Traversing p => Mapping p where

Methods

roam :: ((a -> b) -> s -> t) -> p i a b -> p i s t

iroam :: ((i -> a -> b) -> s -> t) -> p j a b -> p (i -> j) s t

class Visiting p => Traversing p where

Methods

wander :: (forall f. Applicative f => (a -> f b) -> s -> f t) -> p i a b -> p i s t

iwander :: (forall f. Applicative f => (i -> a -> f b) -> s -> f t) -> p j a b -> p (i -> j) s t

Concrete profunctors

newtype Star f i a b

Needed for traversals.

Constructors

Star

Fields

runStar :: a -> f b

reStar :: Star f i a b -> Star f j a b

Repack Star to change its index type.

newtype Forget r i a b

Needed for getters and folds.

Constructors

Forget

Fields

runForget :: a -> r

reForget :: Forget r i a b -> Forget r j a b

Repack Forget to change its index type.

newtype ForgetM r i a b

Needed for affine folds.

Constructors

ForgetM

Fields

runForgetM :: a -> Maybe r

newtype FunArrow i a b

Needed for setters.

Constructors

FunArrow

Fields

runFunArrow :: a -> b

reFunArrow :: FunArrow i a b -> FunArrow j a b

Repack FunArrow to change its index type.

newtype IxStar f i a b

Needed for indexed traversals.

Constructors

IxStar

Fields

runIxStar :: i -> a -> f b

newtype IxForget r i a b

Needed for indexed folds.

Constructors

IxForget

Fields

runIxForget :: i -> a -> r

newtype IxForgetM r i a b

Needed for indexed affine folds.

Constructors

IxForgetM

Fields

runIxForgetM :: i -> a -> Maybe r

newtype IxFunArrow i a b

Needed for indexed setters.

Constructors

IxFunArrow

Fields

runIxFunArrow :: i -> a -> b

data StarA f i a b

Needed for conversion of affine traversal back to its VL representation.

Constructors

StarA (forall r. r -> f r) (a -> f b)

runStarA :: StarA f i a b -> a -> f b

Unwrap StarA.

data IxStarA f i a b

Needed for conversion of indexed affine traversal back to its VL representation.

Constructors

IxStarA (forall r. r -> f r) (i -> a -> f b)

runIxStarA :: IxStarA f i a b -> i -> a -> f b

Unwrap StarA.

data Exchange a b i s t

Constructors

Exchange (s -> a) (b -> t)

data Store a b i s t

Type to represent the components of a lens.

Constructors

Store (s -> a) (s -> b -> t)

data Market a b i s t

Type to represent the components of a prism.

Constructors

Market (b -> t) (s -> Either t a)

data AffineMarket a b i s t

Type to represent the components of an affine traversal.

Constructors

AffineMarket (s -> b -> t) (s -> Either t a)

newtype Tagged i a b

Tag a value with not one but two phantom type parameters (so that Tagged can be used as an indexed profunctor).

Constructors

Tagged

Fields

unTagged :: b

data Context a b t

Constructors

Context (b -> t) a

Utilities

(#.) :: Coercible b c => (b -> c) -> (a -> b) -> a -> c infixr 9

Composition operator where the first argument must be an identity function up to representational equivalence (e.g. a newtype wrapper or unwrapper), and will be ignored at runtime.

(.#) :: Coercible a b => (b -> c) -> (a -> b) -> a -> c infixl 8

Composition operator where the second argument must be an identity function up to representational equivalence (e.g. a newtype wrapper or unwrapper), and will be ignored at runtime.